+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2023.09.23 v5.10.0
 
 1. new `nextVersion` function, see: [nextVersion](https://github.com/saqqdy/js-cool#nextversion)
-1. new `promisify` function, see: [promisify](https://github.com/saqqdy/js-cool#promisify)
+1. new `promiseFactory` function, see: [promiseFactory](https://github.com/saqqdy/js-cool#promiseFactory)
 
 ## 2023.09.20 v5.9.0